+++ extensions/pinboard/src/addBookmark.tsx
@@ -1,11 +1,39 @@
-import { Form, ActionPanel, Action, showToast, Icon, getSelectedText, Toast, showHUD, popToRoot } from "@raycast/api";
-import { useState, useEffect } from "react";
-import fetch from "node-fetch";
-import { Bookmark, addBookmark } from "./api";
-import he from "he";
+import { Form, ActionPanel, Action, showToast, Icon, getSelectedText, Toast, popToRoot } from "@raycast/api";
+import { FormValidation, useForm } from "@raycast/utils";
+import { useEffect } from "react";
+import { Bookmark, BookmarkFormValues } from "./types";
+import { addBookmark, loadDocumentTitle } from "./api";
+import { isValidURL } from "./utils";
 
 export default function Command() {
-  const [state, setState] = useState<{ url: string; title: string }>({ url: "", title: "" });
+  const { handleSubmit, itemProps, setValue, focus } = useForm<BookmarkFormValues>({
+    async onSubmit(values) {
+      const toast = await showToast({ title: "Pinning bookmark...", style: Toast.Style.Animated });
+
+      try {
+        await addBookmark(values as Bookmark);
+        toast.style = Toast.Style.Success;
+        toast.title = "Successfully added bookmark";
+        popToRoot();
+      } catch (error) {
+        console.error("addBookmark error", error);
+        toast.title = "Could not pin bookmark";
+        toast.message = String(error);
+        toast.style = Toast.Style.Failure;
+      }
+    },
+    validation: {
+      url: (value) => {
+        if (value?.length === 0) {
+          return "The item is required";
+        } else if (value && !isValidURL(value)) {
+          return "Enter a valid URL";
+        }
+      },
+      title: FormValidation.Required,
+    },
+    initialValues: {},
+  });
 
   useEffect(() => {
     (async () => {
@@ -16,46 +44,24 @@ export default function Command() {
           console.log(selectedText, "is not a valid URL");
           return;
         }
+        setValue("url", selectedText);
         try {
           const documentTitle = await loadDocumentTitle(selectedText);
-          setState((oldState) => ({ ...oldState, url: selectedText, title: documentTitle }));
+          if (documentTitle) {
+            setValue("title", documentTitle);
+            focus("tags");
+          } else {
+            focus("title");
+          }
         } catch (error) {
           console.error("Could not load document title", error);
-          setState((oldState) => ({ ...oldState, url: selectedText }));
+          focus("title");
         }
       } catch (error) {
         console.error("Could not get selected text", error);
       }
     })();
   }, []);
-
-  async function handleSubmit(values: Bookmark) {
-    console.log("bookmark", values);
-    const url = values.url.trim();
-    const title = values.title.trim();
-    if (!isValidURL(url) || title.length === 0) {
-      showToast({ title: "Enter a valid URL and title for the bookmark", style: Toast.Style.Failure });
-      return;
-    }
-    const toast = await showToast({ title: "Pinning bookmark...", style: Toast.Style.Animated });
-    try {
-      await addBookmark(values);
-      toast.hide();
-      showHUD("Bookmark pinned!");
-      popToRoot();
-    } catch (error) {
-      console.error("addBookmark error", error);
-      showToast({ title: "Could not pin bookmark", message: String(error), style: Toast.Style.Failure });
-    }
-  }
-
-  function handleURLChange(value: string) {
-    setState((oldState) => ({ ...oldState, url: value }));
-  }
-
-  function handleTitleChange(value: string) {
-    setState((oldState) => ({ ...oldState, title: value }));
-  }
 
   return (
     <Form
@@ -67,45 +73,15 @@ export default function Command() {
       }
     >
       <Form.TextField
-        id="url"
         title="URL"
         placeholder="Enter URL (Tip: Select a URL before opening this form)"
-        value={state.url}
-        onChange={handleURLChange}
+        {...itemProps.url}
       />
-      <Form.TextField
-        id="title"
-        title="Title"
-        placeholder="Enter title"
-        value={state.title}
-        onChange={handleTitleChange}
-      />
+      <Form.TextField title="Title" placeholder="Enter title" {...itemProps.title} />
       <Form.Separator />
-      <Form.TextField id="tags" title="Tags" placeholder="Enter tags (comma-separated)" />
-      <Form.Checkbox id="private" title="" label="Private" storeValue />
-      <Form.Checkbox id="readLater" title="" label="Read Later" storeValue />
+      <Form.TextField title="Tags" placeholder="Enter tags (comma separated)" {...itemProps.tags} />
+      <Form.Checkbox title="" label="Private" storeValue {...itemProps.private} />
+      <Form.Checkbox title="" label="Read Later" storeValue {...itemProps.readLater} />
     </Form>
   );
-}
-
-async function loadDocumentTitle(url: string): Promise<string> {
-  const response = await fetch(url);
-  if (!response.ok) {
-    return Promise.reject(response.statusText);
-  }
-  return extractDocumentTitle(await response.text());
-}
-
-function extractDocumentTitle(document: string): string {
-  const title = document.match(/<title>(.*?)<\/title>/)?.[1] ?? "";
-  return he.decode(title);
-}
-
-function isValidURL(url: string): boolean {
-  try {
-    new URL(url);
-    return true;
-  } catch (err) {
-    return false;
-  }
 }