+++ extensions/pinboard/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [Enhancements] - 2023-02-15
+
+- Added cache
+- Optimized extension
+
 ## [Metadata] - 2022-03-23
 
 - Added new metadata for the store